+++ tests/analysis/test_fitting_models.py
@@ -1,4 +1,5 @@
 """Tests for analysis fitting models"""
+
 import numpy as np
 import pytest