+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - SI utilities - Cast large values for seconds to minutes or hours (!501)
 - Documentation
   - Update copyright notice to 2024. (!506)
+- Analysis - Added analysis class for readout calibration (!474)
 - GitLab
   - Make general Merge Request template the default. (!507)
   - Add documentation redirect instructions to Release merge request template. (!507)